Moved action creation for dark current correction from BasicScalerChannelDetector to within the CLSSIS3820Scaler class. Created a new action class to perform the dark current correction--AMDoDarkCurrentCorrectionAction/Info. Tested, works.

+++ source/beamline/AMDetector.cpp
@@ -396,7 +396,7 @@ void AMDetector::setAsDarkCurrentCorrection(){
     }
 }
 
-void AMDetector::setLastDarkCurrentCorrectionTime(int lastTime) {
+void AMDetector::setLastDarkCurrentCorrectionTime(double lastTime) {
     if (canDoDarkCurrentCorrection()) {
         lastDarkCurrentCorrectionTime_ = lastTime;
     }
